+++ tiled/client/container.py
@@ -112,21 +112,24 @@ class Container(BaseClient, collections.abc.Mapping, IndexersMixin):
             # In the Python API we encode sorting as (key, direction).
             # This order-based "record" notion does not play well with OpenAPI.
             # In the HTTP API, therefore, we use {"key": key, "direction": direction}.
+            server_sorting = item["attributes"].get("sorting") or []
             self._sorting = [
-                (s["key"], int(s["direction"]))
-                for s in (item["attributes"].get("sorting") or [])
-            ]
-        sorting = sorting or item["attributes"].get("sorting")
-        self._sorting_params = {
-            "sort": ",".join(
-                f"{'-' if item[1] < 0 else ''}{item[0]}" for item in self._sorting
-            )
-        }
-        self._reversed_sorting_params = {
-            "sort": ",".join(
-                f"{'-' if item[1] > 0 else ''}{item[0]}" for item in self._sorting
-            )
-        }
+                (s["key"], int(s["direction"])) for s in server_sorting
+            ] or [
+                ("", 1)
+            ]  # fall back to server default when not specified
+        sorting_list = [
+            f"{'-' if item[1] < 0 else ''}{item[0]}" for item in self._sorting
+        ]
+        reversed_sorting_list = [
+            item[1:] if item.startswith("-") else f"-{item.lstrip('+')}"
+            for item in sorting_list
+        ]
+        self._sorting_params = {"sort": sorting_list} if "".join(sorting_list) else {}
+        self._reversed_sorting_params = (
+            {"sort": reversed_sorting_list} if "".join(reversed_sorting_list) else {}
+        )
+
         super().__init__(
             context=context,
             item=item,